Move pointer handling to component
